+++ bot/requirements_gcp.txt
@@ -1,0 +1,2 @@
+python-telegram-bot>=21.0.1,<22.0.0
+functions-framework>=3.5.0,<3.6.0
